+++ butler_offline_selenium_tests/add_kategorie_on_page_test.py
@@ -1,4 +1,4 @@
-from butler_offline_selenium_tests.SeleniumTest import SeleniumTestClass
+from butler_offline_selenium_tests.selenium_test import SeleniumTestClass
 from butler_offline_selenium_tests.page.util import enter_test_mode, fill_element
 from selenium.webdriver.common.by import By
 import time
