+++ butler_offline_selenium_tests/konfigurations_test.py
@@ -1,4 +1,4 @@
-from butler_offline_selenium_tests.SeleniumTest import SeleniumTestClass
+from butler_offline_selenium_tests.selenium_test import SeleniumTestClass
 from butler_offline_selenium_tests.page.util import enter_test_mode
 from butler_offline_selenium_tests.page.gemeinsam.gemeinsam_add import GemeinsamAdd
 from butler_offline_selenium_tests.page.core.configuration import Configuration
